add RouteInfoManager
